+++ sdk/core/azure-core/src/io/body_stream.cpp
@@ -75,11 +75,10 @@ std::vector<uint8_t> BodyStream::ReadToEnd(Context const& context)
   }
 }
 
-int64_t MemoryBodyStream::OnRead(uint8_t* buffer, size_t count, Context const& context)
+size_t MemoryBodyStream::OnRead(uint8_t* buffer, size_t count, Context const& context)
 {
   (void)context;
-  int64_t copy_length = std::min(
-      static_cast<int64_t>(count), static_cast<int64_t>(this->m_length - this->m_offset));
+  size_t copy_length = std::min(count, this->m_length - this->m_offset);
   // Copy what's left or just the count
   std::memcpy(buffer, this->m_data + m_offset, static_cast<size_t>(copy_length));
   // move position
@@ -176,7 +175,7 @@ FileBodyStream::~FileBodyStream()
 #endif
 }
 
-int64_t FileBodyStream::OnRead(uint8_t* buffer, size_t count, Azure::Core::Context const& context)
+size_t FileBodyStream::OnRead(uint8_t* buffer, size_t count, Azure::Core::Context const& context)
 {
   return m_randomAccessFileBodyStream->Read(buffer, count, context);
 }
